requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
defusedxml==0.7.1
Django==4.2.4
django-cors-headers==4.2.0
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-xml==2.0.0
Faker==4.14.2
Pillow==10.0.0
python-dateutil==2.8.1
pytz==2020.4
redis==5.0.0
six==1.15.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.7.1
tzdata==2023.3
validate-docbr==1.8.2